refactor: extract route modules from server.ts monolith (ARC-2)
Extract 11 route module files from the 2,720-line server.ts monolith:
- routes/context.ts: RouteContext interface, guards, helpers
- routes/health.ts: health, prometheus, alerts, handshake, swarm
- routes/auth.ts: auth verify, API keys CRUD, SSE token
- routes/audit.ts: audit log, global metrics, diagnostics
- routes/sessions.ts: session CRUD, listing, batch delete
- routes/session-actions.ts: send, read, answer, interrupt, kill, etc.
- routes/session-data.ts: transcript, summary, screenshot, tools, SSE
- routes/events.ts: global SSE stream
- routes/templates.ts: template CRUD
- routes/pipelines.ts: batch create, pipeline CRUD
- routes/index.ts: barrel export

server.ts reduced from 2,720 to 1,130 lines (58% reduction).
Guards and helpers parameterized instead of closing over module vars.

Closes #1695
